feat: [CDS-43512]: Make capacity context independent
