+++ 125-cd-nextgen/src/main/java/io/harness/cdng/common/capacity/CapacitySpec.java
@@ -5,14 +5,19 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.cdng.spot.elastigroup.deploy;
+package io.harness.cdng.common.capacity;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 
+import static com.fasterxml.jackson.annotation.JsonTypeInfo.As.EXTERNAL_PROPERTY;
+import static com.fasterxml.jackson.annotation.JsonTypeInfo.Id.NAME;
+
 import io.harness.annotations.dev.OwnedBy;
 
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
 @OwnedBy(CDP)
-public interface CapacitySpecType {
-  String COUNT = "Count";
-  String PERCENTAGE = "Percentage";
+@JsonTypeInfo(use = NAME, property = "type", include = EXTERNAL_PROPERTY, visible = true)
+public interface CapacitySpec {
+  String getType();
 }
