+++ 125-cd-nextgen/src/main/java/io/harness/cdng/common/capacity/PercentageCapacitySpec.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-package io.harness.cdng.spot.elastigroup.deploy;
+package io.harness.cdng.common.capacity;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 import static io.harness.beans.SwaggerConstants.INTEGER_CLASSPATH;
@@ -20,6 +20,7 @@ import io.harness.yaml.YamlSchemaTypes;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import io.swagger.annotations.ApiModelProperty;
+import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import lombok.Builder;
 import lombok.Data;
@@ -27,10 +28,10 @@ import lombok.Getter;
 
 @Data
 @Builder
-@JsonTypeName(CapacitySpecType.COUNT)
+@JsonTypeName(CapacitySpecType.PERCENTAGE)
 @OwnedBy(CDP)
-@RecasterAlias("io.harness.cdng.spot.elastigroup.deploy.CountCapacitySpec")
-public class CountCapacitySpec implements CapacitySpec {
+@RecasterAlias("io.harness.cdng.common.capacity.PercentageCapacitySpec")
+public class PercentageCapacitySpec implements CapacitySpec {
   @JsonProperty(YamlNode.UUID_FIELD_NAME)
   @Getter(onMethod_ = { @ApiModelProperty(hidden = true) })
   @ApiModelProperty(hidden = true)
@@ -38,12 +39,13 @@ public class CountCapacitySpec implements CapacitySpec {
 
   @ApiModelProperty(dataType = INTEGER_CLASSPATH)
   @YamlSchemaTypes(value = {expression})
-  @JsonProperty("count")
+  @JsonProperty("percentage")
   @Min(0)
-  ParameterField<Integer> count;
+  @Max(100)
+  ParameterField<Integer> percentage;
 
   @Override
   public String getType() {
-    return CapacitySpecType.COUNT;
+    return CapacitySpecType.PERCENTAGE;
   }
 }
