+++ 125-cd-nextgen/src/main/java/io/harness/cdng/spot/elastigroup/deploy/ElastigroupDeployStepParameters.java
@@ -11,6 +11,7 @@ import static io.harness.annotations.dev.HarnessTeam.CDP;
 
 import io.harness.annotation.RecasterAlias;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.common.capacity.Capacity;
 import io.harness.plancreator.steps.TaskSelectorYaml;
 import io.harness.pms.yaml.ParameterField;
