+++ 125-cd-nextgen/src/test/java/io/harness/cdng/spot/elastigroup/deploy/ElastigroupDeployStepInfoTest.java
@@ -15,6 +15,9 @@ import io.harness.CategoryTest;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.category.element.UnitTests;
+import io.harness.cdng.common.capacity.Capacity;
+import io.harness.cdng.common.capacity.CapacitySpecType;
+import io.harness.cdng.common.capacity.CountCapacitySpec;
 import io.harness.plancreator.steps.TaskSelectorYaml;
 import io.harness.plancreator.steps.common.SpecParameters;
 import io.harness.pms.execution.OrchestrationFacilitatorType;
