+++ 125-cd-nextgen/src/test/resources/recaster/alias/alias.txt
@@ -537,8 +537,8 @@ io.harness.cdng.infra.yaml.ElastigroupInfrastructure
 io.harness.cdng.elastigroup.output.ElastigroupConfigurationOutput
 io.harness.cdng.freeze.FreezeOutcome
 io.harness.cdng.spot.elastigroup.deploy.ElastigroupDeployStepParameters
-io.harness.cdng.spot.elastigroup.deploy.PercentageCapacitySpec
-io.harness.cdng.spot.elastigroup.deploy.CountCapacitySpec
+io.harness.cdng.common.capacity.PercentageCapacitySpec
+io.harness.cdng.common.capacity.CountCapacitySpec
 io.harness.cdng.spot.elastigroup.deploy.ElastigroupDeployStepInfo
 io.harness.cdng.spot.elastigroup.deploy.ElastigroupDeployStepNode
-io.harness.cdng.spot.elastigroup.deploy.Capacity
+io.harness.cdng.common.capacity.Capacity
